+++ Practice/src/practice07/PTra07_06.java
@@ -10,22 +10,18 @@ public class PTra07_06 {
 	public static void main(String[] args) {
 
 		// ★ ※2次元配列を用いています
-		int[][] array = {
-				{500,40,6},
-				{70,8,90,1000}
-		};
+		int[][] array = { { 500, 40, 6 }, { 70, 8, 90, 1000 } };
 
 		// ★ 配列arrayに格納されている全ての数字の平均を出力してください
 		int sum = 0;
 		int length = 0;
 		for (int i = 0; i < array.length; i++) {
-
-			for (int value: array[i]) {
+			for (int value : array[i]) {
 				sum += value;
 			}
 			length += array[i].length;
 		}
-		System.out.println(sum / length);
+		System.out.println((double) sum / length);
 
 	}
 }
